pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "monitor-slurm"
version = "1.0.0"
description = "Monitor SLURM jobs and receive notifications via Telegram."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
    "python-telegram-bot>=13.0,<21.0",
    "python-dotenv>=0.15.0"
]

[project.scripts]
monitor-slurm = "monitor_slurm.cli:main"

[tool.setuptools.packages.find]
where = ["."]

